Add "object id" based equality checking and hashing to resource objects

+++ web_glitz/src/extensions/color_buffer_float.rs
@@ -60,7 +60,7 @@ use crate::runtime::Connection;
 /// See the [color_buffer_float] module documentation for details.
 #[derive(Clone, Debug)]
 pub struct Extension {
-    context_id: usize,
+    context_id: u64,
 }
 
 impl Extension {
@@ -106,7 +106,7 @@ impl Extension {
 }
 
 impl super::Extension for Extension {
-    fn try_init(connection: &mut Connection, context_id: usize) -> Option<Self> {
+    fn try_init(connection: &mut Connection, context_id: u64) -> Option<Self> {
         let (gl, _) = unsafe { connection.unpack() };
 
         gl.get_extension("EXT_color_buffer_float")
